Refactor Folder logic into FolderManager
