+++ app/src/main/kotlin/com/x8bit/bitwarden/data/vault/repository/VaultRepositoryImpl.kt
@@ -10,12 +10,9 @@ import com.bitwarden.core.data.repository.util.map
 import com.bitwarden.core.data.repository.util.mapNullable
 import com.bitwarden.core.data.repository.util.updateToPendingOrLoading
 import com.bitwarden.core.data.util.asFailure
-import com.bitwarden.core.data.util.flatMap
 import com.bitwarden.data.manager.DispatcherManager
 import com.bitwarden.exporters.ExportFormat
 import com.bitwarden.fido.Fido2CredentialAutofillView
-import com.bitwarden.network.model.UpdateFolderResponseJson
-import com.bitwarden.network.service.FolderService
 import com.bitwarden.network.util.isNoConnectionError
 import com.bitwarden.sdk.Fido2CredentialStore
 import com.bitwarden.send.SendView
@@ -34,14 +31,13 @@ import com.x8bit.bitwarden.data.platform.error.MissingPropertyException
 import com.x8bit.bitwarden.data.platform.error.NoActiveUserException
 import com.x8bit.bitwarden.data.platform.manager.DatabaseSchemeManager
 import com.x8bit.bitwarden.data.platform.manager.PushManager
-import com.x8bit.bitwarden.data.platform.manager.model.SyncFolderDeleteData
-import com.x8bit.bitwarden.data.platform.manager.model.SyncFolderUpsertData
 import com.x8bit.bitwarden.data.platform.repository.util.observeWhenSubscribedAndLoggedIn
 import com.x8bit.bitwarden.data.platform.repository.util.observeWhenSubscribedAndUnlocked
 import com.x8bit.bitwarden.data.vault.datasource.disk.VaultDiskSource
 import com.x8bit.bitwarden.data.vault.datasource.sdk.VaultSdkSource
 import com.x8bit.bitwarden.data.vault.manager.CipherManager
 import com.x8bit.bitwarden.data.vault.manager.CredentialExchangeImportManager
+import com.x8bit.bitwarden.data.vault.manager.FolderManager
 import com.x8bit.bitwarden.data.vault.manager.SendManager
 import com.x8bit.bitwarden.data.vault.manager.TotpCodeManager
 import com.x8bit.bitwarden.data.vault.manager.VaultLockManager
@@ -50,21 +46,17 @@ import com.x8bit.bitwarden.data.vault.manager.model.GetCipherResult
 import com.x8bit.bitwarden.data.vault.manager.model.ImportCxfPayloadResult
 import com.x8bit.bitwarden.data.vault.manager.model.SyncVaultDataResult
 import com.x8bit.bitwarden.data.vault.manager.model.VerificationCodeItem
-import com.x8bit.bitwarden.data.vault.repository.model.CreateFolderResult
-import com.x8bit.bitwarden.data.vault.repository.model.DeleteFolderResult
 import com.x8bit.bitwarden.data.vault.repository.model.DomainsData
 import com.x8bit.bitwarden.data.vault.repository.model.ExportVaultDataResult
 import com.x8bit.bitwarden.data.vault.repository.model.GenerateTotpResult
 import com.x8bit.bitwarden.data.vault.repository.model.ImportCredentialsResult
 import com.x8bit.bitwarden.data.vault.repository.model.SendData
 import com.x8bit.bitwarden.data.vault.repository.model.TotpCodeResult
-import com.x8bit.bitwarden.data.vault.repository.model.UpdateFolderResult
 import com.x8bit.bitwarden.data.vault.repository.model.VaultData
 import com.x8bit.bitwarden.data.vault.repository.model.VaultUnlockResult
 import com.x8bit.bitwarden.data.vault.repository.util.sortAlphabetically
 import com.x8bit.bitwarden.data.vault.repository.util.sortAlphabeticallyByTypeAndOrganization
 import com.x8bit.bitwarden.data.vault.repository.util.toDomainsData
-import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedNetworkFolder
 import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCipher
 import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCipherList
 import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCollectionList
@@ -87,7 +79,6 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filter
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
@@ -116,12 +107,12 @@ private const val STOP_TIMEOUT_DELAY_MS: Long = 1000L
  */
 @Suppress("TooManyFunctions", "LongParameterList", "LargeClass")
 class VaultRepositoryImpl(
-    private val folderService: FolderService,
     private val vaultDiskSource: VaultDiskSource,
     private val vaultSdkSource: VaultSdkSource,
     private val authDiskSource: AuthDiskSource,
     private val settingsDiskSource: SettingsDiskSource,
     private val cipherManager: CipherManager,
+    private val folderManager: FolderManager,
     private val sendManager: SendManager,
     private val vaultLockManager: VaultLockManager,
     private val totpCodeManager: TotpCodeManager,
@@ -133,6 +124,7 @@ class VaultRepositoryImpl(
     private val credentialExchangeImportManager: CredentialExchangeImportManager,
 ) : VaultRepository,
     CipherManager by cipherManager,
+    FolderManager by folderManager,
     SendManager by sendManager,
     VaultLockManager by vaultLockManager {
 
@@ -280,16 +272,6 @@ class VaultRepositoryImpl(
             .fullSyncFlow
             .onEach { sync(forced = false) }
             .launchIn(unconfinedScope)
-
-        pushManager
-            .syncFolderDeleteFlow
-            .onEach(::deleteFolder)
-            .launchIn(unconfinedScope)
-
-        pushManager
-            .syncFolderUpsertFlow
-            .onEach(::syncFolderIfNecessary)
-            .launchIn(ioScope)
 
         databaseSchemeManager
             .databaseSchemeChangeFlow
@@ -645,106 +627,6 @@ class VaultRepositoryImpl(
             )
     }
 
-    override suspend fun createFolder(folderView: FolderView): CreateFolderResult {
-        val userId = activeUserId
-            ?: return CreateFolderResult.Error(error = NoActiveUserException())
-        return vaultSdkSource
-            .encryptFolder(
-                userId = userId,
-                folder = folderView,
-            )
-            .flatMap { folder ->
-                folderService
-                    .createFolder(
-                        body = folder.toEncryptedNetworkFolder(),
-                    )
-            }
-            .onSuccess { vaultDiskSource.saveFolder(userId = userId, folder = it) }
-            .flatMap { vaultSdkSource.decryptFolder(userId, it.toEncryptedSdkFolder()) }
-            .fold(
-                onSuccess = { CreateFolderResult.Success(folderView = it) },
-                onFailure = { CreateFolderResult.Error(error = it) },
-            )
-    }
-
-    override suspend fun updateFolder(
-        folderId: String,
-        folderView: FolderView,
-    ): UpdateFolderResult {
-        val userId = activeUserId ?: return UpdateFolderResult.Error(
-            errorMessage = null,
-            error = NoActiveUserException(),
-        )
-        return vaultSdkSource
-            .encryptFolder(
-                userId = userId,
-                folder = folderView,
-            )
-            .flatMap { folder ->
-                folderService
-                    .updateFolder(
-                        folderId = folder.id.toString(),
-                        body = folder.toEncryptedNetworkFolder(),
-                    )
-            }
-            .fold(
-                onSuccess = { response ->
-                    when (response) {
-                        is UpdateFolderResponseJson.Success -> {
-                            vaultDiskSource.saveFolder(userId, response.folder)
-                            vaultSdkSource
-                                .decryptFolder(
-                                    userId,
-                                    response.folder.toEncryptedSdkFolder(),
-                                )
-                                .fold(
-                                    onSuccess = { UpdateFolderResult.Success(it) },
-                                    onFailure = {
-                                        UpdateFolderResult.Error(
-                                            errorMessage = null,
-                                            error = it,
-                                        )
-                                    },
-                                )
-                        }
-
-                        is UpdateFolderResponseJson.Invalid -> {
-                            UpdateFolderResult.Error(
-                                errorMessage = response.message,
-                                error = null,
-                            )
-                        }
-                    }
-                },
-                onFailure = { UpdateFolderResult.Error(it.message, error = it) },
-            )
-    }
-
-    override suspend fun deleteFolder(folderId: String): DeleteFolderResult {
-        val userId = activeUserId
-            ?: return DeleteFolderResult.Error(error = NoActiveUserException())
-        return folderService
-            .deleteFolder(
-                folderId = folderId,
-            )
-            .onSuccess {
-                clearFolderIdFromCiphers(folderId, userId)
-                vaultDiskSource.deleteFolder(userId, folderId)
-            }
-            .fold(
-                onSuccess = { DeleteFolderResult.Success },
-                onFailure = { DeleteFolderResult.Error(error = it) },
-            )
-    }
-
-    private suspend fun clearFolderIdFromCiphers(folderId: String, userId: String) {
-        vaultDiskSource.getCiphers(userId = userId).forEach {
-            if (it.folderId == folderId) {
-                vaultDiskSource.saveCipher(userId = userId, cipher = it.copy(folderId = null))
-            }
-        }
-    }
-
     override suspend fun exportVaultDataToString(
         format: ExportFormat,
         restrictedTypes: List<CipherType>,
@@ -1048,48 +930,6 @@ class VaultRepositoryImpl(
                 settingsDiskSource.getLastSyncTime(userId = userId) == null
             }
             ?: DataState.Loading
-
-    //region Push notification helpers
-
-    /**
-     * Deletes the folder specified by [syncFolderDeleteData] from disk.
-     */
-    private suspend fun deleteFolder(syncFolderDeleteData: SyncFolderDeleteData) {
-        clearFolderIdFromCiphers(
-            folderId = syncFolderDeleteData.folderId,
-            userId = syncFolderDeleteData.userId,
-        )
-        vaultDiskSource.deleteFolder(
-            folderId = syncFolderDeleteData.folderId,
-            userId = syncFolderDeleteData.userId,
-        )
-    }
-
-    /**
-     * Syncs an individual folder contained in [syncFolderUpsertData] to disk if certain criteria
-     * are met.
-     */
-    private suspend fun syncFolderIfNecessary(syncFolderUpsertData: SyncFolderUpsertData) {
-        val userId = activeUserId ?: return
-        val folderId = syncFolderUpsertData.folderId
-        val isUpdate = syncFolderUpsertData.isUpdate
-        val revisionDate = syncFolderUpsertData.revisionDate
-        val localFolder = vaultDiskSource
-            .getFolders(userId = userId)
-            .first()
-            .find { it.id == folderId }
-        val isValidCreate = !isUpdate && localFolder == null
-        val isValidUpdate = isUpdate &&
-            localFolder != null &&
-            localFolder.revisionDate.toEpochSecond() < revisionDate.toEpochSecond()
-
-        if (!isValidCreate && !isValidUpdate) return
-
-        folderService
-            .getFolder(folderId = folderId)
-            .onSuccess { vaultDiskSource.saveFolder(userId = userId, folder = it) }
-    }
-    //endregion Push Notification helpers
 
     private suspend fun syncInternal(
         userId: String,
